+++ Assets/Scripts/Weapon/WeaponDefinition.cs
@@ -3,8 +3,8 @@
     public WeaponType Type;
     public ushort MaxAmmo;
     public ushort BulletsPerMagazine;
-    public int DamagePerBullet;
-    public int HeadShotDamagePerBullet;
+    public float DamagePerBullet;
+    public float HeadShotDamagePerBullet;
     public float ReloadTime;
     public float ShotInterval;
     public bool IsAutomatic;
